+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* QueueConfigReader: added the ability to specify default values for settings:
+    * database table 
+    * ExtSettings
 * Generic DAO implementation moved from db-queue-spring to new module db-queue-jdbc
     * db-queue-jdbc may be used without spring. You need to implement interface JdbcOperationsSubset
     * Spring specific remains in db-queue-spring. Contains JdbcOperationsSubset and SpringDatabaseAccessLayer